move info command to archiver.info
